Improve project structure by using alternative way of routing (#2)
* Start restructuring project. Remove /base/models

* Add suppliers 'app'

* Add products 'app'

* Add login 'app'

* Move routes out of basepagemaker

* Fix imports

* Update routing

* Add common folder

* Move content to common_folder

* Fix imports

* Fix imports

* Move file from libs to specific helper file

* Rename base to warehouse

* Move product templates into app folder

* Move templates to login app

* Add templates for supplier app

* Move ean and gs1 routes to products

* Remove duplicate readsession. Move requestUserSettings to long app.

* Update urls.py

* Move file
