--- base/config.ini
@@ -1,4 +1,0 @@
-[mysql]
-user = warehouse
-password = warehouse
-database = warehouse
