--- base/model/__init__.py
@@ -1,3 +1,0 @@
-from .basemodel import *
-from .product import *
-from .supplier import *
