+++ warehouse/basepages.py
@@ -5,18 +5,13 @@ import locale
 
 # standard modules
 import time
-from io import StringIO
 
 # uweb modules
 import uweb3
 from uweb3.libs import mail
 
-# project modules
-from base import model
-from base.decorators import NotExistsErrorCatcher
-from base.pages import products, supplier
-
-from .helpers import PagedResult
+from warehouse.common import model as common_model
+from warehouse.login import model as login_model
 
 
 def CentRound(monies):
@@ -26,7 +21,8 @@ def CentRound(monies):
 
 
 class PageMaker(
-    uweb3.DebuggingPageMaker, uweb3.LoginMixin, products.PageMaker, supplier.PageMaker
+    uweb3.DebuggingPageMaker,
+    uweb3.LoginMixin,
 ):
     """Holds all the request handlers for the application"""
 
@@ -64,7 +60,7 @@ class PageMaker(
 
     def _PreRequest(self):
         list(
-            model.User.List(self.connection)
+            login_model.User.List(self.connection)
         )  # XXX: mysql connection bugged. Remove this
         if self.config.Read():
             try:
@@ -88,184 +84,28 @@ class PageMaker(
         )
         return response
 
-    @uweb3.decorators.TemplateParser("login.html")
-    def RequestLogin(self, url=None):
-        """Please login"""
-        if self.user:
-            return self.RequestIndex()
-        if not url and "url" in self.get:
-            url = self.get.getfirst("url")
-        return {"url": url}
-
-    @uweb3.decorators.checkxsrf
-    @uweb3.decorators.TemplateParser("logout.html")
-    def RequestLogout(self):
-        """Handles logouts"""
-        message = "You where already logged out."
-        if self.user:
-            message = ""
-            if "action" in self.post:
-                session = model.Session(self.connection)
-                session.Delete()
-                message = "Logged out."
-        return {"message": message}
-
-    @uweb3.decorators.checkxsrf
-    def HandleLogin(self):
-        """Handles a username/password combo post."""
-        if self.user or "email" not in self.post or "password" not in self.post:
-            return self.RequestIndex()
-        url = (
-            self.post.getfirst("url", None)
-            if self.post.getfirst("url", "").startswith("/")
-            else "/"
-        )
-        try:
-            self._user = model.User.FromLogin(
-                self.connection,
-                self.post.getfirst("email"),
-                self.post.getfirst("password"),
-            )
-            model.Session.Create(self.connection, int(self.user), path="/")
-            print("login successful.", self.post.getfirst("email"))
-            # redirect 303 to make sure we GET the next page, not post again to avoid leaking login details.
-            return self.req.Redirect(url, httpcode=303)
-        except uweb3.uweb3.model.NotExistError as error:
-            self.parser.RegisterTag("loginerror", "%s" % error)
-            print("login failed.", self.post.getfirst("email"))
-        return self.RequestLogin(url)
-
-    @uweb3.decorators.checkxsrf
-    def RequestResetPassword(self, email=None, resethash=None):
-        """Handles the post for the reset password."""
-        message = None
-        error = False
-        if not email and not resethash:
-            try:
-                user = model.User.FromEmail(
-                    self.connection, self.post.getfirst("email", "")
-                )
-            except uweb3.uweb3.model.NotExistError:
-                error = True
-                if self.debug:
-                    print(
-                        "Password reset request for unknown user %s:"
-                        % self.post.getfirst("email", "")
-                    )
-            if not error:
-                resethash = user.PasswordResetHash()
-                content = self.parser.Parse(
-                    "email/resetpass.txt",
-                    email=user["email"],
-                    host=self.options["general"]["host"],
-                    resethash=resethash,
-                )
-                try:
-                    with mail.MailSender(
-                        local_hostname=self.options["general"]["host"]
-                    ) as send_mail:
-                        send_mail.Text(user["email"], "CMS password reset", content)
-                except mail.SMTPConnectError:
-                    if not self.debug:
-                        return self.Error(
-                            "Mail could not be send due to server error, please contact support."
-                        )
-                if self.debug:
-                    print("Password reset for %s:" % user["email"], content)
-
-            message = "If that was an email address that we know, a mail with reset instructions will be in your mailbox soon."
-            return self.parser.Parse("reset.html", message=message)
-        try:
-            user = model.User.FromEmail(self.connection, email)
-        except uweb3.uweb3.model.NotExistError:
-            return self.parser.Parse(
-                "reset.html", message="Sorry, that's not the right reset code."
-            )
-        if resethash != user.PasswordResetHash():
-            return self.parser.Parse(
-                "reset.html", message="Sorry, that's not the right reset code."
-            )
-
-        if "password" in self.post:
-            if self.post.getfirst("password") == self.post.getfirst(
-                "password_confirm", ""
-            ):
-                try:
-                    user.UpdatePassword(self.post.getfirst("password", ""))
-                except ValueError:
-                    return self.parser.Parse(
-                        "reset.html",
-                        message="Password too short, 8 characters minimal.",
-                    )
-                model.Session.Create(self.connection, int(user), path="/")
-                self._user = user
-                return self.parser.Parse(
-                    "reset.html",
-                    message="Your password has been updated, and you are logged in.",
-                )
-            else:
-                return self.parser.Parse(
-                    "reset.html", message="The passwords don't match."
-                )
-        return self.parser.Parse(
-            "resetform.html", resethash=resethash, resetuser=user, message=""
-        )
-
     def _ReadSession(self):
         """Attempts to read the session for this user from his session cookie"""
         try:
-            user = model.Session(self.connection)
+            user = login_model.Session(self.connection)
         except Exception:
             raise ValueError("Session cookie invalid")
         try:
-            user = model.User.FromPrimary(self.connection, int(str(user)))
+            user = login_model.User.FromPrimary(self.connection, int(str(user)))
         except uweb3.model.NotExistError:
-            return None
+            raise ValueError("User not valid")
         if user["active"] != "true":
             raise ValueError("User not active, session invalid")
         return user
 
-    @uweb3.decorators.checkxsrf
-    @uweb3.decorators.TemplateParser("setup.html")
-    def RequestSetup(self):
-        """Allows the user to setup various fields, and create an admin user.
-
-        If these fields are already filled out, this page will not function any
-        longer.
-        """
-        if not model.User.IsFirstUser(self.connection):
-            return self.RequestLogin()
-
-        if (
-            "email" in self.post
-            and "password" in self.post
-            and "password_confirm" in self.post
-            and self.post.getfirst("password") == self.post.getfirst("password_confirm")
-        ):
-            try:
-                user = model.User.Create(
-                    self.connection,
-                    {
-                        "ID": 1,
-                        "email": self.post.getfirst("email"),
-                        "active": "true",
-                        "password": self.post.getfirst("password"),
-                    },
-                    generate_password_hash=True,
-                )
-            except ValueError:
-                return {"error": "Password too short, 8 characters minimal."}
-
-            model.Session.Create(self.connection, int(user), path="/")
-
-            self.config.Create("general", "host", self.post.getfirst("hostname"))
-            self.config.Create(
-                "general", "locale", self.post.getfirst("locale", "en_GB")
-            )
-            model.Session.Create(self.connection, int(user), path="/")
-            return self.req.Redirect("/", httpcode=301)
-        if self.post:
-            return {"error": "Not all fields are properly filled out."}
+    @uweb3.decorators.TemplateParser("login.html")
+    def RequestLogin(self, url=None):
+        """Please login"""
+        if self.user:
+            return uweb3.Redirect("/products")
+        if not url and "url" in self.get:
+            url = self.get.getfirst("url")
+        return {"url": url}
 
     @uweb3.decorators.loggedin
     @uweb3.decorators.checkxsrf
@@ -275,7 +115,7 @@ class PageMaker(
         if self.user["ID"] != 1:
             return self.req.Redirect("/")
 
-        currentusers = list(model.User.List(self.connection))
+        currentusers = list(login_model.User.List(self.connection))
         if self.post:
             values = {}
             for key in (
@@ -335,7 +175,7 @@ class PageMaker(
         # handle User creation
         if "useremail" in self.post and "new" in values["useremail"]:
             try:
-                newuser = model.User.Create(
+                newuser = login_model.User.Create(
                     self.connection,
                     {
                         "email": values["useremail"].get("new", "").strip(),
@@ -352,7 +192,7 @@ class PageMaker(
                         "users": users,
                     }
                 users.append(newuser)
-            except model.InvalidNameError:
+            except common_model.InvalidNameError:
                 return {
                     "usererror": "Provide a valid email address for the new user.",
                     "users": users,
@@ -384,43 +224,10 @@ class PageMaker(
 
     @uweb3.decorators.loggedin
     @uweb3.decorators.checkxsrf
-    @uweb3.decorators.TemplateParser("usersettings.html")
-    def RequestUserSettings(self):
-        """Returns the user settings page."""
-        # handle password change
-        if "password" in self.post or "password_confirm" in self.post:
-            password = self.post.getfirst("password", "")
-            password_confirm = self.post.getfirst("password_confirm", "")
-            if password != password_confirm:
-                return {"error": "Passwords do not match, try again."}
-            try:
-                self.user.UpdatePassword(password)
-            except ValueError:
-                return {"error": "Passwords too short."}
-            else:
-                content = self.parser.Parse(
-                    "email/updateuser.txt", email=self.user["email"]
-                )
-                try:
-                    with mail.MailSender(
-                        local_hostname=self.options["general"]["host"]
-                    ) as send_mail:
-                        send_mail.Text(
-                            self.user["email"], "Warehouse account change", content
-                        )
-                except mail.SMTPConnectError:
-                    if not self.debug:
-                        return self.Error(
-                            "Mail could not be send due to server error, please contact support."
-                        )
-            return {"succes": "Password has been updated."}
-
-    @uweb3.decorators.loggedin
-    @uweb3.decorators.checkxsrf
     @uweb3.decorators.TemplateParser("apisettings.html")
     def RequestApiSettings(self):
         """Returns the api settings page."""
-        currentkeys = list(model.Apiuser.List(self.connection))
+        currentkeys = list(login_model.Apiuser.List(self.connection))
 
         # handle api key updates
         keys = []
@@ -478,11 +285,11 @@ class PageMaker(
         # handle new api key creation
         if "new_name" in self.post and len(self.post.getfirst("new_name")) > 0:
             try:
-                newkey = model.Apiuser.Create(
+                newkey = login_model.Apiuser.Create(
                     self.connection, {"name": self.post.getfirst("new_name")}
                 )
                 keys.append(newkey)
-            except model.InvalidNameError:
+            except common_model.InvalidNameError:
                 return {
                     "keys": keys,
                     "apierror": "Provide a valid name for the new API key.",
@@ -501,36 +308,49 @@ class PageMaker(
     @uweb3.decorators.loggedin
     def RequestIndex(self):
         """Returns the homepage"""
-        return self.RequestProducts()
+        return self.req.Redirect("/products")
 
-    @uweb3.decorators.loggedin
-    @uweb3.decorators.TemplateParser("gs1.html")
-    def RequestGS1(self):
-        """Returns the gs1 page"""
-        products = PagedResult(
-            self.pagesize,
-            self.get.getfirst("page", 1),
-            model.Product.List,
-            self.connection,
-            {"conditions": ["gs1 is not null"], "order": [("gs1", False)]},
-        )
-        return {"products": products}
+    @uweb3.decorators.checkxsrf
+    @uweb3.decorators.TemplateParser("setup.html")
+    def RequestSetup(self):
+        """Allows the user to setup various fields, and create an admin user.
 
-    @uweb3.decorators.loggedin
-    @uweb3.decorators.TemplateParser("ean.html")
-    def RequestEAN(self):
-        """Returns the EAN page"""
-        products = PagedResult(
-            self.pagesize,
-            self.get.getfirst("page", 1),
-            model.Product.List,
-            self.connection,
-            {
-                "conditions": ["(gs1 is not null or ean is not null)"],
-                "order": [("ean", False)],
-            },
-        )
-        return {"products": products}
+        If these fields are already filled out, this page will not function any
+        longer.
+        """
+        if not login_model.User.IsFirstUser(self.connection):
+            return self.RequestLogin()
+
+        if (
+            "email" in self.post
+            and "password" in self.post
+            and "password_confirm" in self.post
+            and self.post.getfirst("password") == self.post.getfirst("password_confirm")
+        ):
+            try:
+                user = login_model.User.Create(
+                    self.connection,
+                    {
+                        "ID": 1,
+                        "email": self.post.getfirst("email"),
+                        "active": "true",
+                        "password": self.post.getfirst("password"),
+                    },
+                    generate_password_hash=True,
+                )
+            except ValueError:
+                return {"error": "Password too short, 8 characters minimal."}
+
+            login_model.Session.Create(self.connection, int(user), path="/")
+
+            self.config.Create("general", "host", self.post.getfirst("hostname"))
+            self.config.Create(
+                "general", "locale", self.post.getfirst("locale", "en_GB")
+            )
+            login_model.Session.Create(self.connection, int(user), path="/")
+            return self.req.Redirect("/", httpcode=301)
+        if self.post:
+            return {"error": "Not all fields are properly filled out."}
 
     def XSRFInvalidToken(self):
         """Show that the users XSRF token is b0rked"""
