+++ warehouse/common/helpers.py
