+++ warehouse/config.ini
@@ -1,0 +1,15 @@
+[mysql]
+user = warehouse
+password = warehouse
+database = warehouse
+
+[development]
+port = 8002
+
+[signedCookie]
+secret = xWpGikKrpUS+o0HmZbWXwm+ijdZ1yyFMljeCp5D8eVN4baO7h+qLndyFmSSOcx1GUlbo/VryG5KyKMUhR4HsW+rt9yoM7pWY2OhPlqaeg3CGYBhw9fVihnG9/zyEBjdRhQzhqkqz8Ju5jQDjTJhs31vVPxB/0/gjsqZATVPH7/U=
+
+[general]
+host = 127.0.0.1
+locale = en_GB
+port = 8002
