+++ warehouse/products/model.py
@@ -7,14 +7,12 @@ __version__ = "1.0"
 # standard modules
 import datetime
 import math
-import re
 
 import pytz
 from uweb3 import model
 
-from base.model import basemodel
-
-__all__ = ["Product", "Stock", "Productpart"]
+from warehouse.common import model as common_model
+from warehouse.login import model as login_model
 
 
 class Product(model.Record):
@@ -28,7 +26,10 @@ class Product(model.Record):
     def List(cls, connection, conditions=[], *args, **kwargs):
         """Returns the Products filtered on not deleted"""
         return super().List(
-            connection, conditions=[basemodel.NOTDELETED] + conditions, *args, **kwargs
+            connection,
+            conditions=[common_model.NOTDELETED] + conditions,
+            *args,
+            **kwargs
         )
 
     @classmethod
@@ -77,7 +78,8 @@ class Product(model.Record):
         with connection as cursor:
             product = cursor.Select(
                 table=cls.TableName(),
-                conditions=["name=%s" % safe_name, basemodel.NOTDELETED] + conditions,
+                conditions=["name=%s" % safe_name, common_model.NOTDELETED]
+                + conditions,
             )
         if not product:
             raise cls.NotExistError("There is no product with common name %r" % name)
@@ -100,7 +102,7 @@ class Product(model.Record):
         if not self["sku"]:  # set empty string to None for key contraints
             self["sku"] = None
         if not self["name"]:
-            raise basemodel.InvalidNameError("Provide a valid name")
+            raise common_model.InvalidNameError("Provide a valid name")
 
     def _PreSave(self, cursor):
         super()._PreSave(cursor)
@@ -114,7 +116,7 @@ class Product(model.Record):
         if not self["sku"]:  # set empty string to None for key contraints
             self["sku"] = None
         if not self["name"]:
-            raise basemodel.InvalidNameError("Provide a valid name")
+            raise common_model.InvalidNameError("Provide a valid name")
 
     @property
     def parts(self):
@@ -191,23 +193,23 @@ class Product(model.Record):
         if amount > 0:
             possiblestock = self.possiblestock
             if not possiblestock["available"] and not possiblestock["limitedby"]:
-                raise basemodel.AssemblyError(
+                raise common_model.AssemblyError(
                     "Cannot assemble this product, is not an assembled product."
                 )
             if not possiblestock["available"] or possiblestock["available"] < amount:
-                raise basemodel.AssemblyError(
+                raise common_model.AssemblyError(
                     "Cannot assemble this product, not enough parts. Limited by: %s"
                     % possiblestock["limitedby"]["part"]["name"]
                 )
             parts = possiblestock["parts"]
         elif amount < 0:
             if self.currentstock < abs(amount):
-                raise basemodel.AssemblyError(
+                raise common_model.AssemblyError(
                     "Cannot Disassemble this product, not enough stock available."
                 )
             parts = list(self.parts)
             if parts == 0:
-                raise basemodel.AssemblyError(
+                raise common_model.AssemblyError(
                     "Cannot Disassemble this product, is not an assembled product."
                 )
         return parts
