+++ warehouse/static/favicon.ico
