+++ warehouse/suppliers/templates/parts
@@ -1,0 +1,1 @@
+/home/stef/devel/uweb3warehouse/warehouse/templates/parts
