+++ warehouse/templates/admin.html
