+++ warehouse/templates/apisettings.html
