+++ warehouse/templates/email/newuser.txt
